Fix argument type for isprint().
Some platforms define isprint() as a macro which must have an integer
argument instead of char.  Add explicit cast to make everyone happy.

+++ src/find_string.c
@@ -76,7 +76,7 @@ nvm_string_find(const char* blob, size_t size, uint8_t min_length)
 		--printable;
 	    }
 	    printable = 0;
-	} else if (isprint(*c)) {	//printable character, might be part of string
+	} else if (isprint((int) *c)) {	//printable character, might be part of string
 	    if (printable < UINT8_MAX) ++printable;
 #ifdef DEBUG
 	    printf("'%c'", *c);
